+++ backend/canisters/community/api/src/lifecycle/post_upgrade.rs
@@ -1,8 +1,10 @@
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
-use types::BuildVersion;
+use std::collections::HashSet;
+use types::{BuildVersion, UserId};
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
+    pub deleted_users: HashSet<UserId>,
     pub wasm_version: BuildVersion,
 }