+++ backend/canisters/community/impl/src/model/invited_users.rs
@@ -53,4 +53,8 @@ impl InvitedUsers {
     pub fn len(&self) -> usize {
         self.users.len()
     }
+
+    pub fn iter(&self) -> impl Iterator<Item = (&UserId, &UserInvitation)> {
+        self.users.iter()
+    }
 }