+++ backend/canisters/local_group_index/impl/src/guards.rs
@@ -8,6 +8,14 @@ pub fn caller_is_group_index_canister() -> Result<(), String> {
     }
 }
 
+pub fn caller_is_user_index_canister() -> Result<(), String> {
+    if read_state(|state| state.is_caller_user_index_canister()) {
+        Ok(())
+    } else {
+        Err("Caller is not the user_index canister".to_owned())
+    }
+}
+
 pub fn caller_is_local_group_or_community_canister() -> Result<(), String> {
     if read_state(|state| state.is_caller_local_group_canister() || state.is_caller_local_community_canister()) {
         Ok(())