+++ backend/canisters/local_group_index/impl/src/updates/mod.rs
@@ -7,6 +7,7 @@ pub mod c2c_set_community_upgrade_concurrency;
 pub mod c2c_set_group_upgrade_concurrency;
 pub mod c2c_set_max_concurrent_community_upgrades;
 pub mod c2c_set_max_concurrent_group_upgrades;
+mod c2c_sync_deleted_users;
 pub mod c2c_trigger_upgrade;
 pub mod c2c_upgrade_community_canister_wasm;
 pub mod c2c_upgrade_group_canister_wasm;